requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
channels==4.3.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
Deprecated==1.2.18
Django==5.2.6
django-cors-headers==4.7.0
django-modeltranslation==0.19.16
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
drf-spectacular==0.28.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lxml==6.0.0
markdown2==2.5.4
MarkupSafe==3.0.2
mysqlclient==2.2.7
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
svix==1.67.0
types-Deprecated==1.2.15.20250304
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
wrapt==1.17.2
django-mcp-server
django-oauth-toolkit==3.0.1
oauthlib==3.3.1
jwcrypto==1.5.6
celery==5.4.0
redis==5.2.0
docker==7.1.0
